Fix src/arch.h, src/cli/proot.c, src/extension/{fake_id0/fake_id0.c,link2symlink/link2symlink.c}, src/path/temp.c, src/syscall/{exit.c,rlimit.c,socket.c}, src/tracee/{event.c,seccomp.c}

+++ src/syscall/rlimit.c
@@ -88,7 +88,7 @@ int translate_setrlimit_exit(const Tracee *tracee, bool is_prlimit)
 		/* Convert this special value from 32-bit to 64-bit,
 		 * if needed.  */
 		if (is_32on64_mode(tracee) && tracee_stack_limit == (uint32_t) -1)
-			tracee_stack_limit = RLIM_INFINITY;
+			tracee_stack_limit = (uint32_t)RLIM_INFINITY;
 	}
 	if (errno != 0)
 		return -errno;
@@ -112,6 +112,6 @@ int translate_setrlimit_exit(const Tracee *tracee, bool is_prlimit)
 		VERBOSE(tracee, 1, "can't set stack limit.");
 	return 0; /* Not fatal.  */
 
-	VERBOSE(tracee, 1, "stack soft limit increased to %ld bytes", proot_stack.rlim_cur);
+	VERBOSE(tracee, 1, "stack soft limit increased to %ld bytes", (long int)proot_stack.rlim_cur);
 	return 0;
 }
